+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 - LibUI general control and window support
 - LibUI listener support
+- LibUI property support
 - Support examples/basic_window.rb